added payPage (uncommented)
Not really done with this. I haven't been able to test it since I can't create assignments yet. But as far as I can tell, this should function. I'm not really sure about the format I have right now. I'm considering having the tableview show more information about the assignment (guide/hotel/etc.). I'm also not entirely sure how I want to handle the status of the assignment. For now, if the assignment cannot be paid for, it should throw an error and raise a popup explaining why. But I'm considering just not showing assignments that have been paid for.  Or maybe disabling selection on rows for items that have been paid for already.