added comments to pay page
still not that happy with the way assignment statuses were handled, but this will do until I care more about it.

TODO: For some reason, control nodes will not deselect when the user clicks elsewhere in pages: PayPage, SetupNMCPage, AssignmentsPage, and TripsPage. This does not appear in MembersPage. I can't for the life of me figure out why this is.

It's actually somewhat of a significant issue. The TableView object in AssignmentsPage, if selected, ends up bleeding over to other pages on my laptop. I.e. the blue outline it gets partially shows up when you click on other pages.